Updated CraftBukkit to build #2805

+++ src/no/runsafe/framework/internal/wrapper/entity/BukkitLivingEntity.java
@@ -26,7 +26,7 @@ public abstract class BukkitLivingEntity extends RunsafeEntity
 		return entity;
 	}
 
-	public int getHealth()
+	public double getHealth()
 	{
 		return entity.getHealth();
 	}
